UXD-304 Fix: seperated variations
